PoC: Creation of an Application folder on the user's Solid Pod
This proof of concept shows how to create an application folder on the
user's Solid Pod and shows how to create a simple file in it.
This commit is a great start to understand how to handle files and
folders on a Pod. With that done we should be confident in the
implementation of forms because when the app folder is created, we just
need to create one file per form.
Note: this implementation is just a proof of concept and needs some
changes before using it in a real application.